Clean up names in ref matchers
