+++ lib/lexers/lexRangeA1.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-mixed-operators */
 import { REF_RANGE, REF_BEAM, REF_TERNARY, MAX_COLS, MAX_ROWS } from '../constants.js';
 import { advRangeOp } from './advRangeOp.js';
-import { isNextCharVerboten } from './isNextCharVerboten.js';
+import { canEndRange, canEndPartialRange } from './canEndRange.js';
 
 function advA1Col (str, pos) {
   // [A-Z]{1,3}
@@ -71,7 +71,7 @@ export function lexRangeA1 (str, pos, options) {
     const top = advA1Row(str, p);
     p += top;
     const op = advRangeOp(str, p);
-    const pre_op = p;
+    const preOp = p;
     if (op) {
       p += op;
       right = advA1Col(str, p);
@@ -79,24 +79,24 @@ export function lexRangeA1 (str, pos, options) {
       bottom = advA1Row(str, p);
       p += bottom;
       if (top && bottom && right) {
-        if (!isNextCharVerboten(str, p) && options.mergeRefs) {
+        if (canEndRange(str, p) && options.mergeRefs) {
           return { type: REF_RANGE, value: str.slice(pos, p) };
         }
       }
       else if (!top && !bottom) {
-        if (!isNextCharVerboten(str, p)) {
+        if (canEndRange(str, p)) {
           return { type: REF_BEAM, value: str.slice(pos, p) };
         }
       }
       else if (options.allowTernary) {
-        if (!isNextCharVerboten(str, p, true)) {
+        if (canEndPartialRange(str, p)) {
           return { type: REF_TERNARY, value: str.slice(pos, p) };
         }
       }
     }
     // LT  : this is A1
-    if (top && !isNextCharVerboten(str, pre_op)) {
-      return { type: REF_RANGE, value: str.slice(pos, pre_op) };
+    if (top && canEndRange(str, preOp)) {
+      return { type: REF_RANGE, value: str.slice(pos, preOp) };
     }
   }
   else {
@@ -115,12 +115,12 @@ export function lexRangeA1 (str, pos, options) {
         bottom = advA1Row(str, p);
         p += bottom;
         if (right && bottom && options.allowTernary) {
-          if (!isNextCharVerboten(str, p, true)) {
+          if (canEndPartialRange(str, p)) {
             return { type: REF_TERNARY, value: str.slice(pos, p) };
           }
         }
         if (!right && bottom) {
-          if (!isNextCharVerboten(str, p)) {
+          if (canEndRange(str, p)) {
             return { type: REF_BEAM, value: str.slice(pos, p) };
           }
         }
