+++ lib/lexers/lexRangeR1C1.js
@@ -1,20 +1,14 @@
 /* eslint-disable no-mixed-operators */
-import {
-  REF_RANGE,
-  REF_BEAM,
-  REF_TERNARY,
-  MAX_COLS,
-  MAX_ROWS
-} from '../constants.js';
+import { REF_RANGE, REF_BEAM, REF_TERNARY, MAX_COLS, MAX_ROWS } from '../constants.js';
 import { advRangeOp } from './advRangeOp.js';
-import { isNextCharVerboten } from './isNextCharVerboten.js';
+import { canEndRange } from './canEndRange.js';
 
 const BR_OPEN = 91; // [
 const BR_CLOSE = 93; // ]
-const uR = 82;
-const lR = 114;
-const uC = 67;
-const lC = 99;
+const UC_R = 82;
+const LC_R = 114;
+const UC_C = 67;
+const LC_C = 99;
 const PLUS = 43;
 const MINUS = 45;
 
@@ -27,7 +21,7 @@ const MINUS = 45;
 function lexR1C1Part (str, pos, isRow = false) {
   const start = pos;
   const c0 = str.charCodeAt(pos);
-  if ((isRow ? c0 === uR || c0 === lR : c0 === uC || c0 === lC)) {
+  if ((isRow ? c0 === UC_R || c0 === LC_R : c0 === UC_C || c0 === LC_C)) {
     pos++;
     let digits = 0;
     let value = 0;
@@ -101,7 +95,7 @@ export function lexRangeR1C1 (str, pos, options) {
   p += c1;
   if (c1 || r1) {
     const op = advRangeOp(str, p);
-    const pre_op = p;
+    const preOp = p;
     if (op) {
       p += op;
       const r2 = lexR1C1Part(str, p, true); // R1
@@ -119,7 +113,7 @@ export function lexRangeR1C1 (str, pos, options) {
         (r1 && c1 && r2 && !c2) ||
         (r1 && c1 && !r2 && c2)
       ) {
-        if (options.allowTernary && !isNextCharVerboten(str, p)) {
+        if (options.allowTernary && canEndRange(str, p)) {
           return { type: REF_TERNARY, value: str.slice(pos, p) };
         }
       }
@@ -129,7 +123,7 @@ export function lexRangeR1C1 (str, pos, options) {
         (c1 && c2 && !r1 && !r2) ||
         (!c1 && !c2 && r1 && r2)
       ) {
-        if (!isNextCharVerboten(str, p)) {
+        if (canEndRange(str, p)) {
           return { type: REF_BEAM, value: str.slice(pos, p) };
         }
       }
@@ -138,10 +132,10 @@ export function lexRangeR1C1 (str, pos, options) {
     // R1
     // C1
     // R1C1
-    if (!isNextCharVerboten(str, pre_op)) {
+    if (canEndRange(str, preOp)) {
       return {
         type: (r1 && c1) ? REF_RANGE : REF_BEAM,
-        value: str.slice(pos, pre_op)
+        value: str.slice(pos, preOp)
       };
     }
   }
